requirements.txt
dnspython
discord.py
discord-py-slash-command
python-dotenv
bscscan-python